+++ edgy/core/db/querysets/base.py
@@ -60,9 +60,9 @@ def clean_query_kwargs(model: Type["Model"], kwargs: Dict[str, Any]) -> Dict[str
                     except KeyError:
                         model_class = model_class.meta.related_fields[part].related_from
         if field_name in model_class.meta.fields_mapping:
-            new_kwargs.update(model_class.meta.fields_mapping[field_name].clean(key, val))
+            new_kwargs.update(model_class.meta.fields_mapping[field_name].clean(key, val, for_query=True))
         elif field_name in model_class.meta.related_fields:
-            new_kwargs.update(model_class.meta.related_fields[field_name].clean(key, val))
+            new_kwargs.update(model_class.meta.related_fields[field_name].clean(key, val, for_query=True))
         else:
             new_kwargs[key] = val
     assert "pk" not in new_kwargs, "pk should be already parsed"
@@ -193,6 +193,8 @@ class BaseQuerySet(
                     model_class = related_field.related_from
                     foreign_key = related_field.foreign_key
                     inverse = True
+                if foreign_key.is_cross_db:
+                    raise NotImplementedError("We cannot cross databases yet, this feature is planned")
                 table = model_class.table
                 select_from = sqlalchemy.sql.join(  # type: ignore
                     select_from,